Generate the qtl matrix file
This file present for each marker and each traits the LOD value found by MapQTL.
The approach relies on consistent maps accross the different mqo file and in case of inconsistency
an exception is raised. Since this exception should not block the whole pipeline, the exception
is catched, a warning is given and the analysis continues.