Add converter to a MapChart compatible map.
With this commit we generate a .map file which can be read into MapChart
for further visualisation by biologists.

The QTL interval for MapChart is calculated using a LOD 2 interval if the
LOD score of the peak allows it. Otherwise, it takes the distance between
the LOD of the peak and the LOD Threshold.

The commits also fix the tests to remove the .map file generated when
running the tests.