Expand the interface
Add a valid_file method used when the user provides directly a single file
instead of a whole directory.

This change implies that convert_inputfiles and get_session_identifiers
can receive either a folder or a single input file.

Update current plugins to check that the inputs is correct and not overlapping