Implement adding flanking marker to the list of QTLs retrieved.
With this change the list of QTLs found (more precisely the file 'qtls_with_mk.csv') will
contain not only the closest marker to the peak but also the flanking marker of the LOD2
interval.

The search for the flanking markers has been fixed and now correctly retrieve the next
marker outside of the LOD2 interval.

It should be noted that cXX.posY markers as generated by R/qtl are ignored.

This has changed the MapChart generation as it now also relies on markers from the
genetic map while before it was able to use generated marker and thus had a smaller
interval.